requirements.txt
APScheduler==3.8.1
Django==3.2.9
djangorestframework==3.12.4
mixer==7.2.0
django_apscheduler==0.6.0
psycopg2-binary==2.9.1
dj-database-url==0.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0